Update Java samples
The Java samples have quite a lot of room for improvement.

I changed the following:

 - make the code compile
 - update the readme to point to the latest driver
 - remove one leading tab from every line
 - remove redundant imports
 - use Java 7 try-with-resources
 - release database objects as early as possible
 - propagate exceptions
 - use bind parameters to avoid SQL-injection vulnerabilities when the
   code is copy and pasted and adjusted
  - I had to guess on some of the parameter types since I didn't have a
    schema. I had a look at ql_in-memory_oltp_sample.sql. Executable
    samples with a continuous integration build would be a nice future
    improvement.
